requirements.txt
flask>=3.0.0
python-dotenv>=1.0.1
transformers>=4.41.0
huggingface_hub>=0.23.0
